third_party/blink/web_tests/external/wpt/css/cssom/CSSStyleSheet-constructable-replace-on-regular-sheet-expected.txt

This is a testharness.js-based test.
PASS CSSStyleSheet.replaceSync throws NotAllowedError for non-constructed sheets
PASS CSSStyleSheet.replace returns a rejected promise for non-constructed sheets
FAIL CSSStyleSheet.replace returns a rejected promise for non-constructed sheets that have no associated document assert_equals: old sheet should still apply after replace expected "rgb(0, 255, 0)" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
